+++ scripts/istio_virtualservice_canary_dark.sh
@@ -17,16 +17,26 @@ echo "CLUSTER_NAMESPACE=${CLUSTER_NAMESPACE}"
 
 if [ -z "${VIRTUAL_SERVICE_FILE}" ]; then VIRTUAL_SERVICE_FILE=istio_virtualservice_canary_dark.yaml ; fi
 if [ ! -f ${VIRTUAL_SERVICE_FILE} ]; then
+  if [ -z "${DEPLOYMENT_FILE}" ]; then DEPLOYMENT_FILE=deployment.yml ; fi
+  if [ ! -f ${DEPLOYMENT_FILE} ]; then
+      echo -e "${red}Kubernetes deployment file '${DEPLOYMENT_FILE}' not found${no_color}"
+      exit 1
+  fi
+  # Install 'yq' to process yaml files
+  python -m site &> /dev/null && export PATH="$PATH:`python -m site --user-base`/bin"
+  pip install yq
+  echo -e "Updating $DEPLOYMENT_FILE to represent canary deployment: add label version, modify deployment name"
+  DEPLOYMENT_NAME=$( cat ${DEPLOYMENT_FILE} | yq -r '. | select(.kind=="Deployment") | .metadata.name' ) # read deployment name
   cat > ${VIRTUAL_SERVICE_FILE} << EOF
 apiVersion: networking.istio.io/v1alpha3
 kind: VirtualService
 metadata:
-  name: virtual-service-${IMAGE_NAME}
+  name: virtual-service-${DEPLOYMENT_NAME}
 spec:
   hosts:
     - '*'
   gateways:
-    - gateway-${IMAGE_NAME}
+    - gateway-${DEPLOYMENT_NAME}
   http:
     - match:
         - headers:
@@ -34,16 +44,28 @@ spec:
               regex: '.*Firefox.*'
       route:
         - destination:
-            host: ${IMAGE_NAME}
+            host: ${DEPLOYMENT_NAME}
             subset: canary
     - route:
         - destination:
-            host: ${IMAGE_NAME}
+            host: ${DEPLOYMENT_NAME}
             subset: stable
 EOF
-  #sed -e "s/\${IMAGE_NAME}/${IMAGE_NAME}/g" ${VIRTUAL_SERVICE_FILE}
+  #sed -e "s/\${DEPLOYMENT_NAME}/${DEPLOYMENT_NAME}/g" ${VIRTUAL_SERVICE_FILE}
 fi
 cat ${VIRTUAL_SERVICE_FILE}
 kubectl apply -f ${VIRTUAL_SERVICE_FILE} --namespace ${CLUSTER_NAMESPACE}
 
 kubectl get gateways,destinationrules,virtualservices --namespace ${CLUSTER_NAMESPACE}
+
+echo -e "Gateways, destination rules and virtual services in namespace: ${CLUSTER_NAMESPACE}"
+kubectl get gateway,destinationrule,virtualservice --namespace ${CLUSTER_NAMESPACE}
+
+# echo -e "Installed gateway details:"
+# kubectl get gateway gateway-${DEPLOYMENT_NAME} --namespace ${CLUSTER_NAMESPACE} -o yaml
+
+# echo -e "Installed destination rule details:"
+# kubectl get destinationrule destination-rule-${DEPLOYMENT_NAME} --namespace ${CLUSTER_NAMESPACE} -o yaml
+
+# echo -e "Installed virtual service details:"
+# kubectl get virtualservice virtual-service-${DEPLOYMENT_NAME} --namespace ${CLUSTER_NAMESPACE} -o yaml
